Address an edge case when deactivating entities.
Normally, when a `@:remove` function is called, that means the component has already been removed. However, `deactivate()` calls the listener without removing the component.

I think this is correct, but it has to be clearly documented because it's a bit surprising.

Also, since `destroy()` is removing the components, it should remove them before calling the `@:remove` listener. No need for unusual behavior there!