Improve code completion for systems.
The completion server runs macros in hopes of getting better completion. However, if you're trying to import a class, the macro won't be able to find it, and it will abort, preventing the server from importing that class for you. By skipping the macro (which doesn't add much worth completing), we avoid this.